LayoutTests/inspector/elements/highlight-node.html

<html>
<head>

<style>

body {
    margin: 0;
}
#container {
    width: 400px;
    height: 400px;
    background-color: grey;
}
#inspectedElement {
    margin: 5px;
    border: solid 10px aqua;
    padding: 15px;
    width: 200px;
    height: 200px;
    background-color: blue;
    float: left;
}
#description {
    clear: both;
}

</style>

<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script>

function dumpInspectorHighlightRects()
{
    var rectNames = ["margin", "border", "padding", "content"];
    var rects = window.internals.inspectorHighlightRects(document);
    for (var i = 0; i < rects.length; i++)
    {
        var rectName = (i < rectNames.length ? rectNames[i] : "untitled");
        var rect = rects.item(i);
        var line = rectName + " rect is " + rect.width + " x " + rect.height + " at (" + rect.top + ", " + rect.left + ")<br/>";
        document.getElementById("console").innerHTML += line;
    }
}

function test()
{
    function nodeSelected(node)
    {
        RuntimeAgent.evaluate("dumpInspectorHighlightRects()", InspectorTest.completeTest);
    }

    InspectorTest.selectNodeWithId("inspectedElement", nodeSelected);
}

</script>
</head>

<body onload="runTest()">

<div id="inspectedElement"></div>
<p id="description">This test verifies the position and size of the highlight rectangles overlayed on an inspected div.</p>
<div id="console"></div>

</body>
</html>
